# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation using Swagger"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/api/swagger.json": {
      "get": {
        "summary": "OpenAPI Spec JSON file",
        "description": "Endpoint for a user to extract API Spec in JSON format",
        "operationId": "getSwaggerJson",
        "tags": [
          "Documentation"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Register a new user in the system",
        "operationId": "postRegister",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": null
                }
              }
            }
          },
          "400": {
            "description": "Error registering user"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Login existing user",
        "description": "Login user account with email and password",
        "operationId": "postLogin",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User logged in successfully"
          },
          "400": {
            "description": "Error logging in user"
          }
        }
      }
    },
    "/api/verify": {
      "get": {
        "summary": "Verify token validity",
        "description": "Check if token is expired/if user is still logged in",
        "operationId": "getVerify",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User still logged in with valid token"
          },
          "401": {
            "description": "User session has expired"
          },
          "403": {
            "description": "User session ended due to token error"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "summary": "Logout current user",
        "description": "Clears the authentication cookie, logging out the user.",
        "operationId": "postLogout",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "User not authenticated"
          }
        }
      }
    },
    "/api/dashboard": {
      "post": {
        "summary": "Retrieve user patient data",
        "description": "Queries the user's id (email address) and returns their personal medical information.",
        "operationId": "postDashboard",
        "tags": [
          "User",
          "Patient"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved Patient data"
          },
          "401": {
            "description": "Error retrieving Patient data"
          }
        }
      }
    },
    "/api/update": {
      "patch": {
        "summary": "Update user patient data",
        "description": "Updates the user's patient information with what they entered.",
        "operationId": "postUpdate",
        "tags": [
          "Patient"
        ],
        "responses": {
          "201": {
            "description": "Successfully updated Patient data"
          },
          "401": {
            "description": "Error updating Patient data"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}